input: move tablet tool functions into tablet.c
Just a big move and a rename of the tablet tool destroy handler.
