+++ src/input/meson.build
@@ -2,7 +2,6 @@ labwc_sources += files(
   'cursor.c',
   'tablet.c',
   'tablet-pad.c',
-  'tablet-tool.c',
   'gestures.c',
   'input.c',
   'keyboard.c',
